+++ src/app/mork/celery/emailing_tasks.py
@@ -1,13 +1,16 @@
-"""Mork Celery tasks."""
+"""Mork Celery emailing tasks."""
 
 from datetime import datetime
 from logging import getLogger
 
+from celery import group
 from sqlalchemy import select
 
 from mork.celery.celery_app import app
 from mork.conf import settings
 from mork.database import MorkDB
+from mork.edx.database import OpenEdxDB
+from mork.edx.models.auth import AuthUser
 from mork.exceptions import EmailAlreadySent, EmailSendError
 from mork.mail import send_email
 from mork.models import EmailStatus
@@ -18,13 +21,22 @@ logger = getLogger(__name__)
 @app.task
 def warn_inactive_users():
     """Celery task to warn inactive users by email."""
-    pass
+    db = OpenEdxDB()
 
+    threshold_date = datetime.now() - settings.WARNING_PERIOD
 
-@app.task
-def delete_inactive_users():
-    """Celery task to delete inactive users accounts."""
-    pass
+    total = AuthUser.get_inactive_users_count(db.session, threshold_date)
+    for batch_offset in range(0, total, settings.EDX_QUERY_BATCH_SIZE):
+        inactive_users = AuthUser.get_inactive_users(
+            db.session,
+            threshold_date,
+            offset=batch_offset,
+            limit=settings.EDX_QUERY_BATCH_SIZE,
+        )
+        send_email_group = group(
+            [send_email_task.s(user.email, user.username) for user in inactive_users]
+        )
+        send_email_group.delay()
 
 
 @app.task(
@@ -48,10 +60,10 @@ def send_email_task(self, email_address: str, username: str):
     mark_email_status(email_address)
 
 
-def check_email_already_sent(email_adress: str):
+def check_email_already_sent(email_address: str):
     """Check if an email has already been sent to the user."""
     db = MorkDB()
-    query = select(EmailStatus.email).where(EmailStatus.email == email_adress)
+    query = select(EmailStatus.email).where(EmailStatus.email == email_address)
     result = db.session.execute(query).scalars().first()
     db.session.close()
     return result